backend/requirements.txt
fastapi==0.135.3
starlette==1.0.0
uvicorn[standard]==0.44.0
sqlalchemy[asyncio]==2.0.49
alembic==1.18.4
asyncpg==0.31.0
pydantic==2.12.5
pydantic-settings==2.13.1
authlib==1.6.9
PyJWT[cryptography]==2.12.1
httpx==0.28.1
slowapi==0.1.9
python-multipart==0.0.24
cryptography>=43.0.0
sentry-sdk[fastapi]==2.57.0
jinja2>=3.1.0
Pillow>=11.0.0
pyclamd>=0.4.0
